enhance data type to make possible to support 3D nets

+++ include/caffe/blob.hpp
@@ -356,8 +356,8 @@ class Blob {
   shared_ptr<SyncedMemory> shape_data_;
 #endif
   vector<int> shape_;
-  int count_;
-  int capacity_;
+  long count_;
+  long capacity_;
 
 #ifdef DISTR_WEIGHT_UPDATE
   /* for distributed weight update */
